requirements.txt
web3>=6.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=1.0.0
eth-utils>=2.0.0
eth-abi>=4.0.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0